+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## master (unreleased)
 
+- Prefix this gem log lines with `[JobEnqueueLogger]`
+
 ## 0.2.0 (2023-04-04)
 
 - Support `que`